support no-window env for building (i.e. ssr next.js)

+++ src/configs/storageWrapper.ts
@@ -2,6 +2,7 @@ import { generateUID } from "../utils";
 import type { IStorageWrapper } from "../types";
 
 export class StorageWrapper implements IStorageWrapper {
+    #initialized = false;
     #key: string | undefined = undefined;
     #prefix: string;
     value: string | null | undefined = undefined;
@@ -10,16 +11,15 @@ export class StorageWrapper implements IStorageWrapper {
         this.#prefix = key;
         console.log(`StorageWrapper ${key}`, "constructor");
 
-        // init with prev key/value
-        this.#key = Object.keys(localStorage).find(sk => sk.startsWith(key));
-        console.log(`StorageWrapper ${key}`, "old key", this.#key);
-        if (this.#key) {
-            this.set(localStorage.getItem(this.#key));
-        }
+        this.#init();
     }
 
     set(value: string | null | undefined) {
-        console.log(`StorageWrapper ${this.#key}`, "set", value);
+        console.log(`StorageWrapper ${this.#key || this.#prefix}`, "set", value);
+
+        if (!this.#initialized && !this.#init()) {
+            return false;
+        }
 
         this.delete(); // clear current key
 
@@ -29,25 +29,66 @@ export class StorageWrapper implements IStorageWrapper {
             console.log(`StorageWrapper ${this.#key}`, "set - new key", this.#key);
             localStorage.setItem(this.#key, value);
         }
+
+        return true;
     }
 
     get() {
         this.#validateValue();
-        console.log(`StorageWrapper ${this.#key}`, "get", "value", this.value);
+        console.log(
+            `StorageWrapper ${this.#key || this.#prefix}`,
+            "get",
+            "value",
+            this.value
+        );
         return this.value;
     }
 
     delete() {
-        console.log(`StorageWrapper ${this.#key}`, "delete");
+        console.log(`StorageWrapper ${this.#key || this.#prefix}`, "delete");
+
+        if (!this.#initialized && !this.#init()) {
+            return false;
+        }
 
         this.value = null;
         if (this.#key) localStorage.removeItem(this.#key);
+
+        return true;
     }
 
     #validateValue() {
-        console.log(`StorageWrapper ${this.#key}`, "validateValue", "value", this.value);
+        console.log(
+            `StorageWrapper ${this.#key || this.#prefix}`,
+            "validateValue",
+            "value",
+            this.value
+        );
         if (this.value) {
             this.set(this.value);
         }
+    }
+
+    #init() {
+        try {
+            if (!this.#initialized && typeof window !== "undefined") {
+                // init with prev key/value
+                this.#key = Object.keys(localStorage).find(sk =>
+                    sk.startsWith(this.#prefix)
+                );
+                console.log(`StorageWrapper ${this.#prefix}`, "init", this.#key);
+
+                // set initialized as soon as we managed to extract data
+                // from localStorage, so the `set` call below won't result
+                // in a endless-recursive loop
+                this.#initialized = true;
+                if (this.#key) {
+                    this.set(localStorage.getItem(this.#key));
+                }
+            }
+        } catch (err) {
+            console.warn(err);
+        }
+        return this.#initialized;
     }
 }
